+++ src/lib/components/module/FunctionDetailCard.tsx
@@ -15,8 +15,7 @@ import { LabelText } from "../LabelText";
 import { Tooltip } from "../Tooltip";
 import { AmpEvent, track, trackUseExpand } from "lib/amplitude";
 import { useInternalNavigate, useMobile } from "lib/app-provider";
-import type { IndexedModule } from "lib/services/move/moduleService";
-import type { ExposedFunction } from "lib/types";
+import type { ExposedFunction, IndexedModule } from "lib/types";
 import {
   checkAvailability,
   getFirstQueryParam,