+++ src/lib/components/table/modules/ModulesTableMobileCard.tsx
@@ -1,16 +1,12 @@
-import { Flex, Text } from "@chakra-ui/react";
+import { Flex, Grid, Tag, Text } from "@chakra-ui/react";
 
 import { MobileCardTemplate } from "../MobileCardTemplate";
 import { MobileLabel } from "../MobileLabel";
 import { useInternalNavigate } from "lib/app-provider";
 import { ExplorerLink } from "lib/components/ExplorerLink";
+import { useFormatAddresses } from "lib/hooks/useFormatAddresses";
 import type { ModuleInfo } from "lib/types";
-import {
-  bech32AddressToHex,
-  dateFromNow,
-  formatUTC,
-  unpadHexAddress,
-} from "lib/utils";
+import { dateFromNow, formatUTC } from "lib/utils";
 
 import { ModulePathLink } from "./ModulePathLink";
 
@@ -22,8 +18,9 @@ export const ModulesTableMobileCard = ({
   moduleInfo,
 }: ModulesTableMobileCardProps) => {
   const navigate = useInternalNavigate();
+  const formatAddresses = useFormatAddresses();
+  const { address: creator } = formatAddresses(moduleInfo.address);
 
-  const hex = unpadHexAddress(bech32AddressToHex(moduleInfo.address));
   return (
     <MobileCardTemplate
       onClick={() =>
@@ -31,26 +28,40 @@ export const ModulesTableMobileCard = ({
           pathname: "/modules/[address]/[moduleName]",
           query: {
             address: moduleInfo.address,
-            moduleName: moduleInfo.name,
+            moduleName: moduleInfo.moduleName,
           },
         })
       }
       topContent={
         <Flex direction="column">
           <MobileLabel label="Module Path" />
-          <ModulePathLink hexAddr={hex} moduleName={moduleInfo.name} />
+          <ModulePathLink
+            hexAddr={moduleInfo.address}
+            moduleName={moduleInfo.moduleName}
+          />
         </Flex>
       }
       middleContent={
         <Flex direction="column" gap={3}>
-          <Flex direction="column">
-            <MobileLabel label="owner" />
-            <ExplorerLink
-              value={moduleInfo.address}
-              type="user_address"
-              showCopyOnHover
-            />
-          </Flex>
+          <Grid templateColumns="repeat(2, 1fr)">
+            <Flex direction="column">
+              <MobileLabel label="creator" />
+              <ExplorerLink
+                value={creator}
+                type="user_address"
+                showCopyOnHover
+              />
+            </Flex>
+            <Flex direction="column" gap={1}>
+              <MobileLabel label="Action" />
+              <Tag
+                width="fit-content"
+                variant={moduleInfo.isRepublished ? "primary-light" : "gray"}
+              >
+                {moduleInfo.isRepublished ? "Republish" : "Publish"}
+              </Tag>
+            </Flex>
+          </Grid>
           {moduleInfo.latestUpdated && (
             <Flex direction="column">
               <Text variant="body3">{formatUTC(moduleInfo.latestUpdated)}</Text>