+++ src/lib/pages/interact/component/drawer/ModuleSelectDrawer.tsx
@@ -13,9 +13,8 @@ import { useEffect, useState } from "react";
 import { ModuleEmptyState } from "../common";
 import { useConvertHexAddress } from "lib/app-provider";
 import { CustomIcon } from "lib/components/icon";
-import { useAccountModules } from "lib/services/move/moduleService";
-import type { IndexedModule } from "lib/services/move/moduleService";
-import type { BechAddr, HexAddr, Option } from "lib/types";
+import { useModulesByAddressLcd } from "lib/services/move/moduleService";
+import type { BechAddr, HexAddr, IndexedModule, Option } from "lib/types";
 import { isHexWalletAddress } from "lib/utils";
 
 import { ModuleSelectMainBody } from "./body";
@@ -49,17 +48,13 @@ export const ModuleSelectDrawer = ({
   });
   const [modules, setModules] = useState<IndexedModule[]>();
 
-  const { refetch } = useAccountModules({
+  const { refetch } = useModulesByAddressLcd({
     address: selectedAddress.hex,
-    moduleName: undefined,
-    functionName: undefined,
     options: {
       refetchOnWindowFocus: false,
       enabled: false,
       retry: false,
-      onSuccess: (data) => {
-        if (Array.isArray(data)) setModules(data);
-      },
+      onSuccess: (data) => setModules(data),
     },
   });