+++ src/lib/pages/txs/index.tsx
@@ -2,12 +2,15 @@ import { useRouter } from "next/router";
 import { useEffect } from "react";
 
 import { AmpEvent, track } from "lib/amplitude";
+import { useTierConfig } from "lib/app-provider";
 import PageContainer from "lib/components/PageContainer";
 import { PageHeader } from "lib/components/PageHeader";
 
 import { TxsTable } from "./components/TxsTable";
 
 const Txs = () => {
+  useTierConfig({ minTier: "full" });
+
   const router = useRouter();
 
   useEffect(() => {